2 DC motors
This picture demonstrates my design for the circuit to control 2 DC motors. The circuit contained 2 potentiometer to control each motor separately. This step could be used as initial testing and after analyzing/ obtaining the best angles for each servo, the values can be assigned directly to the move the servos.