Steps:
This picture demonstrates my design for the circuit to control 2 servo motors. The circuit contained 2 potentiometer to control each motor separately. The reason for using potentiometer was because the assigned values can vary depending on the needed movement. This step could be used as initial testing and after analyzing/ obtaining the best angles for each servo where they can function without noisy sound or exceed their limits, the values can be assigned directly to move the servos.