steps:
This picture demonstrates my design for the circuit to control 2 servo motors. The circuit contained 2 potentiometer to control each motor separately. The reason for using potentiometer was because the assigned values can vary depending on the needed/ designed movement. This step could be used as initial testing and after analyzing/ obtaining the best angles for each servo where they can function without noisy sound or exceed their limits, the values can be assigned directly to move the servos. As the power coming from the Arduino is not enough to power both servos, external battery was utilized. An extra component acts as voltage regulator was added for safety purposes and to push 5V to each serv.