Steps:
Controlling DC motor is different from controlling servo motors as the servo's speed and direction of rotation should be controlled. This was achieved with the help of PWM and H-bridge, respectively, which represented the L293D component. The use of potentiometers can be canceled since PWM used to control the speed and it is known that its maximum possible values are 0 to 255. External battery was utilize to make sure the right amount of power will go to the servo and to protect the Arduino.